requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.3
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.4.0
flex==6.14.1
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonpointer==2.2
MarkupSafe==2.0.1
packaging==21.3
platformdirs==2.4.0
protobuf==3.19.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rfc3987==1.3.8
rsa==4.7.2
ruamel.yaml==0.17.18
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
strict-rfc3339==0.7
uritemplate==4.1.1
urllib3==1.26.7
validate-email==1.3
